fix: init config not updating state
