+++ src/actions/ConfigAction.ts
@@ -1,23 +1,8 @@
-import { Action } from 'redux';
-import { Services } from '../types/Service';
-
-export enum ConfigActionType {
-  Save = 'CONFIG_SAVE',
-  Initialize = 'CONFIG_INITIALIZE',
-}
-
-export interface ConfigData {
-  urls: Record<Services, string>;
-  jwtPublicKey: string;
-  googleMapsApiKey: string;
-  logLevelConsole: number;
-  logLevelPersist: number;
-}
-
-export interface ConfigAction extends Action<ConfigActionType> {
-  type: ConfigActionType;
-  payload: ConfigData | null;
-}
+import { ConfigAction, ConfigActionType } from '../definitions/Actions';
+import { Services, Routes } from '../definitions/AppConstants';
+import { ConfigData } from '../definitions/ResponseModels';
+import { getJsonResponse } from '../shared/RequestFactory';
+import { AppThunkAction } from '../store';
 
 export function saveConfig(config: ConfigData): ConfigAction {
   return {
@@ -26,9 +11,8 @@ export function saveConfig(config: ConfigData): ConfigAction {
   };
 }
 
-export function setInitialized(): ConfigAction {
-  return {
-    type: ConfigActionType.Initialize,
-    payload: null,
-  };
-}
+export const fetchConfigAction: AppThunkAction = async (dispatch, _getState) => {
+  const config = await getJsonResponse<ConfigData>(Services.Self, Routes[Services.Self].CONFIG);
+
+  dispatch(saveConfig(config));
+};
