+++ src/actions/LoadingAction.ts
@@ -1,20 +1,5 @@
-import { Action } from 'redux';
-
-export enum LoadingActionType {
-  Start = 'LOADING_START',
-  Report = 'LOADING_REPORT',
-  End = 'LOADING_END',
-}
-
-export interface LoadingInit {
-  isProgressable: boolean;
-  message: string;
-}
-
-export interface LoadingAction extends Action<LoadingActionType> {
-  type: LoadingActionType;
-  payload: number | null | LoadingInit;
-}
+import { LoadingAction, LoadingActionType } from '../definitions/Actions';
+import { Defaults } from '../definitions/AppConstants';
 
 export function loadingStart(isProgressable = false, message = 'Loading... '): LoadingAction {
   return {
@@ -30,6 +15,20 @@ export function loadingReport(percentComplete: number): LoadingAction {
   return {
     type: LoadingActionType.Start,
     payload: percentComplete,
+  };
+}
+
+export function loadingFailed(retryTimeout: number = Defaults.RETRY_TIMEOUT): LoadingAction {
+  return {
+    type: LoadingActionType.Failed,
+    payload: retryTimeout,
+  };
+}
+
+export function loadingClearFailed(): LoadingAction {
+  return {
+    type: LoadingActionType.Failed,
+    payload: null,
   };
 }
