+++ src/actions/LocationAction.ts
@@ -1,19 +1,5 @@
-import { Action } from 'redux';
-import { Coordinates } from '../types/Map';
-
-export enum LocationActionType {
-  LocationUserSet = 'LOCATION_USER_SET',
-  LocationSearchSet = 'LOCATION_SEARCH_SET',
-  LocationSearchClear = 'LOCATION_SEARCH_CLEAR',
-  LocationWatchIdSet = 'LOCATION_WATCH_ID_SET',
-  LocationWatchIdClear = 'LOCATION_WATCH_ID_CLEAR',
-  LocationWatchFailed = 'LOCATION_FAILED',
-}
-
-export interface LocationAction extends Action<LocationActionType> {
-  type: LocationActionType;
-  payload: null | number | Coordinates;
-}
+import { LocationAction, LocationActionType } from '../definitions/Actions';
+import { Coordinates } from '../definitions/Map';
 
 export function setUserLocation(coordinates: Coordinates): LocationAction {
   return {
