+++ src/actions/MapAction.ts
@@ -1,13 +1,6 @@
-import { Action } from 'redux';
-
-export enum MapActionType {
-  Init = 'MAP_INIT',
-}
-
-export interface MapAction extends Action<MapActionType> {
-  type: MapActionType;
-  payload: null;
-}
+import { Loader } from '@googlemaps/js-api-loader';
+import { AppThunkAction } from '../store';
+import { MapAction, MapActionType } from '../definitions/Actions';
 
 export function mapInit(): MapAction {
   return {
@@ -15,3 +8,10 @@ export function mapInit(): MapAction {
     payload: null,
   };
 }
+
+export const fetchMapsApi: AppThunkAction = async (dispatch, getState) => {
+  const apiKey = getState().config.googleMapsApiKey;
+  const mapsLoader = new Loader({ apiKey, libraries: ['places'] });
+  await mapsLoader.load();
+  dispatch(mapInit());
+};
