+++ src/definitions/ComponentModels.ts
@@ -1,8 +1,3 @@
-export interface FormatViewModel {
-  id: string;
-  name: string;
-}
-
 export interface FormatSelection {
   id: string;
   name: string;
