+++ src/definitions/RequestModels.ts
@@ -1,0 +1,5 @@
+export interface StationRequest {
+  lat: number;
+  lng: number;
+  fmt: string[];
+}
