+++ src/shared/State.ts
@@ -1,0 +1,3 @@
+import { LoadingState } from '../definitions/State';
+
+export const isPending = (state: LoadingState) => state.isLoading && state.retryTimeout === null;
