--- src/types/Storage.ts
@@ -1,4 +1,0 @@
-export enum StorageKey {
-  AuthToken = 'AuthToken',
-  FormatState = 'FormatState',
-}
